Move ColorValue and ColorArray into new module
