+++ napari/layers/utils/color_encoding.py
@@ -4,10 +4,11 @@ import numpy as np
 from pydantic import Field, parse_obj_as, validator
 from typing_extensions import Protocol, runtime_checkable
 
+from napari.utils.color import ColorArray, ColorValue
+
 from ...utils import Colormap
 from ...utils.colormaps import ValidColormapArg, ensure_colormap
 from ...utils.colormaps.categorical_colormap import CategoricalColormap
-from ...utils.colormaps.standardize_color import transform_color
 from ...utils.translations import trans
 from .color_transformations import ColorType
 from .style_encoding import (
@@ -16,35 +17,6 @@ from .style_encoding import (
     _DerivedStyleEncoding,
     _ManualStyleEncoding,
 )
-
-
-class ColorValue(np.ndarray):
-    """A 4x1 array that represents one RGBA color value."""
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate_type
-
-    @classmethod
-    def validate_type(cls, val):
-        return transform_color(val)[0]
-
-
-class ColorArray(np.ndarray):
-    """An Nx4 array where each row of N represents one RGBA color value."""
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate_type
-
-    @classmethod
-    def validate_type(cls, val):
-        return (
-            np.empty((0, 4), np.float32)
-            if len(val) == 0
-            else transform_color(val)
-        )
-
 
 """The default color to use, which may also be used a safe fallback color."""
 DEFAULT_COLOR = ColorValue.validate_type('cyan')
